moved common functions for better organization
