+++ projects/reinforcement_learning/foundations/rl/dynamic_programming/policy_methods.py
@@ -1,25 +1,19 @@
-import operator
 from typing import Iterator
 
 import numpy as np
 
 from rl.constants import ACTION, STATE
-from rl.distributions import Categorical, Choose
+from rl.distributions import Choose
+from rl.dynamic_programming import (
+    VALUE_FUNCTION,
+    almost_equal_numpy_arrays,
+    almost_equal_vf_pis,
+    greedy_policy_from_vf,
+)
 from rl.iterate import converged, iterate
 from rl.markov_decision_process import FiniteMarkovDecisionProcess
 from rl.markov_process import FiniteMarkovRewardProcess
 from rl.policy import FiniteDeterministicPolicy, FinitePolicy
-from rl.states import NonTerminal, State
-
-DEFAULT_TOLERANCE = 1e-5
-VALUE_FUNCTION = dict[NonTerminal[STATE], float]
-
-
-def extended_vf(v: VALUE_FUNCTION[STATE], s: State[STATE]) -> float:
-    def _non_terminal_vf(st: NonTerminal[STATE], v=v) -> float:
-        return v[st]
-
-    return s.on_on_terminal(_non_terminal_vf, 0.0)
 
 
 def evaluate_mrp(
@@ -35,14 +29,6 @@ def evaluate_mrp(
     return iterate(step=_update, start=v_0)
 
 
-def almost_equal_numpy_arrays(
-    left: np.ndarray,
-    right: np.ndarray,
-    tolerance: float = DEFAULT_TOLERANCE,
-) -> bool:
-    return max(abs(left - right)) < tolerance
-
-
 def evaluate_mrp_result(
     mrp: FiniteMarkovRewardProcess[STATE],
     gamma: float,
@@ -53,30 +39,6 @@ def evaluate_mrp_result(
     )
 
     return {state: value for state, value in zip(mrp.non_terminal_states, v_star)}
-
-
-def greedy_policy_from_vf(
-    mdp: FiniteMarkovDecisionProcess[STATE, ACTION],
-    vf: VALUE_FUNCTION[STATE],
-    gamma: float,
-) -> FiniteDeterministicPolicy[STATE, ACTION]:
-    greedy_policy: dict[STATE, ACTION] = dict()
-
-    for state in mdp.non_terminal_states:
-        q_values: Iterator[tuple[ACTION, float]] = (
-            (
-                action,
-                mdp.mapping[state][action].expectation(
-                    lambda next_state, reward: reward
-                    + gamma * extended_vf(v=vf, s=next_state)
-                ),
-            )
-            for action in mdp.actions(state)
-        )
-
-        greedy_policy[state.state] = max(q_values, key=operator.itemgetter(1))[0]
-
-    return FiniteDeterministicPolicy(action_for=greedy_policy)
 
 
 def policy_iteration(
@@ -112,16 +74,6 @@ def policy_iteration(
     )
 
     return iterate(step=_update, start=(v_0, policy_0))
-
-
-def almost_equal_vf_pis(
-    left: tuple[VALUE_FUNCTION[STATE], FinitePolicy[STATE, ACTION]],
-    right: tuple[VALUE_FUNCTION[STATE], FinitePolicy[STATE, ACTION]],
-    tolerance: float = DEFAULT_TOLERANCE,
-) -> bool:
-    left_vf, _ = left
-    right_vf, _ = right
-    return max(abs(left_vf[state] - right_vf[state]) for state in left_vf) < tolerance
 
 
 def policy_iteration_result(
